Lightning Interaction Fixes
Now, Lightning collisions check trigger tag. If trigger tag is player, it adds its player controller script to a PlayerController List and for each enables stun on all scripts, start coroutine and then falsifies the stun on every script in the list. This fixed the multi stun perma lock bug